added protection for two simulataneous d/ls writing to same files
git-svn-id: http://svn.vuze.com/public/client@2339 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/disk/file/impl/FMFileImpl.java
@@ -26,6 +26,8 @@ package org.gudy.azureus2.core3.disk.file.impl;
  *
  */
 
+import java.util.*;
+
 import java.io.File;
 import java.io.RandomAccessFile;
 import java.nio.ByteBuffer;
@@ -39,15 +41,33 @@ public abstract class
 FMFileImpl
 	implements FMFile
 {
+	protected static Map			file_map = new HashMap();
+	
+	protected FMFileOwner			owner;
 	protected int					access_mode;
 	protected File					file;
+	protected String				canonical_path;
 	protected RandomAccessFile		raf;
 	
 	protected
 	FMFileImpl(
-		File		_file )
+		FMFileOwner		_owner,
+		File			_file )
+	
+		throws FMFileManagerException
 	{
+		owner	= _owner;
 		file	= _file;
+		
+		try{
+			canonical_path = file.getCanonicalPath();
+			
+			reserveFile();
+			
+		}catch( Throwable e ){
+			
+			throw( new FMFileManagerException( "FMFile::getCanonicalPath: Fails", e ));
+		}
 	}
 
 	public File
@@ -68,15 +88,54 @@ FMFileImpl
 	
 		throws FMFileManagerException
 	{
+		String	new_canonical_path;
+
+		try{
+			new_canonical_path = new_file.getCanonicalPath();
+			
+		}catch( Throwable e ){
+			
+			throw( new FMFileManagerException( "FMFile::getCanonicalPath: Fails", e ));
+		}	
+		
+		if ( new_file.exists()){
+			
+			throw( new FMFileManagerException( "FMFile::moveFile: Fails - file '" + new_canonical_path + "' already exists"));	
+		}
+		
+		boolean	was_open	= raf != null;
+		
 		close();
 		
 		if ( file.renameTo(new_file)) {
 			
-			file	= new_file;
-      
-      openSupport();
+			file			= new_file;
+			canonical_path	= new_canonical_path;
+			
+			reserveFile();
+			
+			openSupport();
 			
 		}else{
+		
+			try{
+				reserveFile();
+				
+			}catch( FMFileManagerException e ){
+				
+				e.printStackTrace();
+			}
+			
+			if ( was_open ){
+				
+				try{
+					openSupport();
+					
+				}catch( FMFileManagerException e){
+					
+					e.printStackTrace();
+				}
+			}
 			
 			throw( new FMFileManagerException( "FMFile::moveFile: Fails"));
 		}	
@@ -116,9 +175,12 @@ FMFileImpl
 	
 		throws FMFileManagerException
 	{
+		if (raf == null){
+			
+			throw new FMFileManagerException( "FMFile::getSize: raf is null" );
+		}
+	      
 		try{
-      if (raf == null) throw new FMFileManagerException( "FMFile::getSize: raf is null" );
-      
 			FileChannel	channel = raf.getChannel();
 			
 			if ( channel.isOpen()){
@@ -154,11 +216,19 @@ FMFileImpl
 	}
 	
 	protected void
-	closeSupport()
+	closeSupport(
+		boolean		explicit )
 	
 		throws FMFileManagerException
 	{
 		if ( raf == null ){
+			
+				// may have previously been implicitly closed, tidy up if required
+			
+			if ( explicit ){
+				
+				releaseFile();
+			}
 			
 			return;
 		}
@@ -173,6 +243,11 @@ FMFileImpl
 		}finally{
 			
 			raf	= null;
+			
+			if ( explicit ){
+				
+				releaseFile();
+			}
 		}
 	}
 	
@@ -183,7 +258,10 @@ FMFileImpl
 	
 		throws FMFileManagerException
 	{
-    if (raf == null) throw new FMFileManagerException( "FMFile::read: raf is null" );
+		if (raf == null){
+			
+			throw new FMFileManagerException( "FMFile::read: raf is null" );
+		}
     
 		FileChannel fc = raf.getChannel();
     		
@@ -217,7 +295,10 @@ FMFileImpl
 	
 		throws FMFileManagerException
 	{
-    if (raf == null) throw new FMFileManagerException( "FMFile::write: raf is null" );
+		if (raf == null){
+			
+			throw new FMFileManagerException( "FMFile::write: raf is null" );
+		}
     
 		FileChannel fc = raf.getChannel();
     
@@ -242,5 +323,38 @@ FMFileImpl
 			
 			throw( new FMFileManagerException( "FMFile::write: Fails", e ));
 		}		
+	}
+	
+	protected void
+	reserveFile()
+	
+		throws FMFileManagerException
+	{
+		synchronized( file_map ){
+			
+			// System.out.println( "FMFile::reserveFile:" + canonical_path + "("+ owner.getName() + ")" );
+			
+			FMFileOwner	existing_owner = (FMFileOwner)file_map.get(canonical_path);
+			
+			if ( existing_owner == null ){
+				
+				file_map.put( canonical_path, owner );
+				
+			}else if ( !existing_owner.getName().equals( owner.getName())){
+				
+				throw( new FMFileManagerException( "File '"+canonical_path+"' is in use by '" + existing_owner.getName()+"'"));
+			}
+		}
+	}
+	
+	protected void
+	releaseFile()
+	{
+		synchronized( file_map ){
+			
+			// System.out.println( "FMFile::releaseFile:" + canonical_path );
+					
+			file_map.remove( canonical_path );
+		}
 	}
 }
